Implement JPA Auditing for Entity Tracking

+++ pos-system/src/main/java/com/lifepill/possystem/dto/requestDTO/ItemSaveRequestCategoryDTO.java
@@ -38,7 +38,7 @@ public class ItemSaveRequestCategoryDTO {
 
         private long categoryId;
 
-        private int supplierId;
+        private long supplierId;
 
 //        private String categoryName;
 //        private String categoryDescription;
